Extract DeprecatedCast
